+++ gnssrefl/rinex2snr.py
@@ -1294,3 +1294,20 @@ def get_local_rinexfile(rfile,localpath2):
                 allgood = True
 
     return allgood
+
+def print_archives():
+    """
+    feeble attempt to print list of archives to screen ...
+
+    """
+    archive_list_rinex3 = ['unavco', 'cddis', 'bev', 'bkg', 'ga', 'epn', 'bfg','sonel','all','unavco2','nrcan','gfz','ignes']
+    archive_list = ['sopac', 'unavco', 'sonel',  'nz', 'ga', 'bkg', 'jeff',
+                    'ngs', 'nrcan', 'special', 'bev', 'jp', 'all','unavco2','cddis']
+    w = 'RINEX 3 archives '
+    for i in range(0,len(archive_list_rinex3)):
+        w = w + archive_list_rinex3[i] + ' '
+    w = w + '\n RINEX 2.11 archives '
+    for i in range(0,len(archive_list)):
+        w = w + archive_list[i] + ' '
+    msg = w
+    return msg